--- apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/branding/resources/page.tsx
@@ -1,5 +1,0 @@
-import { ProgramResourcesPageClient } from "./page-client";
-
-export default function ProgramResources() {
-  return <ProgramResourcesPageClient />;
-}
