+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/resources/program-help-and-support.tsx
@@ -14,7 +14,7 @@ import { SettingsRow } from "../program-settings-row";
 
 type FormData = Pick<ProgramProps, "supportEmail" | "helpUrl" | "termsUrl">;
 
-export function ProgramCommunication() {
+export function ProgramHelpAndSupport() {
   const { program } = useProgram();
   const { id: workspaceId } = useWorkspace();
