+++ apps/web/ui/layout/sidebar/app-sidebar-nav.tsx
@@ -18,9 +18,9 @@ import {
   Globe,
   InvoiceDollar,
   Key,
+  LifeRing,
   LinesY as LinesYStatic,
   MoneyBills2,
-  PaperPlane,
   Receipt2,
   ShieldCheck,
   ShieldKeyhole,
@@ -267,9 +267,9 @@ const NAV_AREAS: SidebarNavAreas<SidebarNavData> = {
             href: `/${slug}/program/link-settings`,
           },
           {
-            name: "Communication",
-            icon: PaperPlane,
-            href: `/${slug}/program/communication`,
+            name: "Resources",
+            icon: LifeRing,
+            href: `/${slug}/program/resources`,
           },
         ],
       },
